NEP_QGEN/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
pypdf2==3.0.1
pillow==10.1.0
pytesseract==0.3.10
google-generativeai==0.3.2
reportlab==4.0.7
python-docx==1.1.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
pydantic==2.5.0
sqlalchemy==2.0.23
